Rename files and folders for consistency
