+++ src/scripts/components/animation-widget/gif-panel-grid.js
