+++ src/scripts/components/animation-widget/loop-button.js
