+++ src/scripts/components/animation-widget/play-button.js
