+++ src/scripts/components/dateline/dateline.js
