+++ src/scripts/components/dateline/text.js
