+++ src/scripts/components/range-selection/dragger-range.js
