+++ src/scripts/components/range-selection/dragger.js
