+++ src/scripts/components/share/links.js
