+++ src/scripts/components/util/checkbox.js
