+++ src/scripts/index.js
@@ -1,13 +1,13 @@
-export {default as AnimationWidget} from './components/animationwidget/wv.animation.widget.js';
-export {default as RangeSelector} from './components/rangeselection/wv.timeline.rangeselection.js';
-export {default as DateSelector} from './components/dateselector/wv.dateselector.js';
-export {default as ToolTip} from './components/tooltip/wv.tooltip.js';
-export {default as Share} from './components/share/wv.share.js';
-export {default as Utils} from './components/util/wv.utils.js';
-export {default as GA} from './components/util/wv.googleAnalytics.js';
-export {default as DateLine} from './components/dateline/wv.line.js';
-export {default as LineText} from './components/dateline/wv.linetext.js';
-export {default as LayerList} from './components/layer/wv.layer.list.js';
-export {default as LayerRow} from './components/layer/wv.layer.row.js';
-export {default as ImageResSelection} from './components/imagepanel/select.js';
-export {default as GifResSelection} from './components/animationwidget/gifpanel.js';
+export {default as AnimationWidget} from './components/animation-widget/animation-widget.js';
+export {default as RangeSelector} from './components/range-selection/range-selection.js';
+export {default as DateSelector} from './components/date-selector/date-selector.js';
+export {default as ToolTip} from './components/tooltip/tooltip.js';
+export {default as Share} from './components/share/share.js';
+export {default as Util} from './components/util/util.js';
+export {default as GA} from './components/util/google-analytics.js';
+export {default as DateLine} from './components/dateline/dateline.js';
+export {default as LineText} from './components/dateline/text.js';
+export {default as LayerList} from './components/layer/list.js';
+export {default as LayerRow} from './components/layer/row.js';
+export {default as ImageResSelection} from './components/image-panel/select.js';
+export {default as GifResSelection} from './components/animation-widget/gif-panel.js';
